change form to router and implementation of login
